Update store and test-utils, moving store logic to its own file with initial testing setup for store
