+++ src/store/store.ts
@@ -1,0 +1,11 @@
+import { rootReducer } from "../reducers/index";
+import { configureStore } from "@reduxjs/toolkit";
+
+export const setupStore = () =>
+  configureStore({
+    reducer: rootReducer,
+  });
+
+export type RootState = ReturnType<typeof rootReducer>;
+export type AppStore = ReturnType<typeof setupStore>;
+export type AppDispatch = AppStore["dispatch"];
